+++ test/components/utils/test_misc.py
@@ -34,7 +34,7 @@ def test_generate_colors():
     expected_output = [Color('#3b4cc0'), Color('#dcdddd'), Color('#b40426')]
 
     expected_input = pd.Series([1.0, 0.5, 0.0])
-    computed_output = generate_colors(expected_input, min=0, max=1)
+    computed_output = generate_colors(expected_input, min=0, max=1, cmap_name='coolwarm_r')
 
     assert computed_output == expected_output